+++ src/Peer.cpp
@@ -82,7 +82,7 @@ std::chrono::steady_clock::time_point Peer::createdAt() const {
 bool Peer::RecvAll(SOCKET sock, char* buffer, int length) {
 	size_t total = 0;
 	while (total < length) {
-		size_t received = recv(sock, buffer + total, length - total, 0);
+		size_t received = recv(sock, buffer + total, (int)(length - total), 0);
 		if (received == 0) {
 			return false;
 		}
@@ -94,7 +94,7 @@ bool Peer::RecvAll(SOCKET sock, char* buffer, int length) {
 bool Peer::SendAll(SOCKET sock, const char* buffer, int length) {
 	size_t total = 0;
 	while (total < length) {
-		size_t sent = send(sock, buffer + total, length - total, 0);
+		size_t sent = send(sock, buffer + total, (int)(length - total), 0);
 		if (sent == 0) {
 			return false;
 		}
